lib modules: defer widget updates while collapsed
Utility modules which respond to the 'collection changed' signal may
end up performing considerable work updating their widgets with large
image libraries (100k+), causing noticable UI lag.  All of this work
is unnecessary when the module is collapsed, as the updates can't be
seen by the user.

This patch adds a new signal that is raised when a module is expanded,
which allows a module to defer widget updates until it is expanded by
the user.  This is applied to the Collections module (which can take
multiple seconds if the active rule is date-time or geotag), as well as
Select, Styles, Tagging, and Metadata.