skip mouse_moved actions for disabled modules
This reduces overhead and thus lag/stuttering when moving the mouse
across the center image in darkroom (especially for the liquify
module, which needs to compute the distortion map *every* time).

It also fixes a subtle UI bug -- when modules have a drawn mask and
the shapes display is enabled, but the module itself is disabled, the
mouse still interacts with the (invisible) drawn shapes.  This leads
to surprises such as a scroll or drag resizing/moving a shape rather
than zooming/panning the image.